// src/util/types/UserType.ts
export interface UserType {
  bio?: string;
  coverImage?: string;
  createdAt: string;
  email?: string;
  emailVerified?: string;
  hasNotification?: string;
  hashedPassword?: string;
  id: string;
  image?: string;
  name?: string;
  profileImage?: string;
  updatedAt: string;
  username: string;
}